+++ webhooks/src/lib/config.ts
@@ -38,6 +38,10 @@ export type SDKOptions = {
    */
   serverURL?: string | undefined;
   /**
+   * Allows overriding the default user agent used by the SDK
+   */
+  userAgent?: string | undefined;
+  /**
    * Allows overriding the default retry config used by the SDK
    */
   retryConfig?: RetryConfig;
@@ -65,8 +69,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.3.5",
-  genVersion: "2.605.6",
+  sdkVersion: "0.4.0",
+  genVersion: "2.625.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.3.5 2.605.6 1.0.0 @avalanche-sdk/webhooks",
+    "speakeasy-sdk/typescript 0.4.0 2.625.0 1.0.0 @avalanche-sdk/webhooks",
 } as const;